Create packages of scenarios to contain the scenario title
It's easier to see the exercises if the title is in the package name.
Rename all world setup text files to WorldSetup.txt.

Update build.xml to handle packages. Because in greenfoot projects we
can't have packages the package names and some import statements have to
be removed by the ant build.